requirements.txt
geopandas==0.12.2
numpy==1.23.5
pandas==1.5.2
Requests==2.30.0
scikit_learn==1.2.0
Shapely==1.8.5.post1
tqdm==4.64.1
pygeos
rtree
torch_geometric
xgboost
torch-geometric-temporal
networkx
pyarrow
statsmodels
torchvision
statsmodels
seaborn